backend/requirements.txt
# =============================================================
#  RL Platform v2.0 — Backend Requirements  (ALL VERSIONS PINNED)
#  SETUP INSIDE VIRTUAL ENVIRONMENT:
#    python3 -m venv venv && source venv/bin/activate
#    pip install -r requirements.txt
# =============================================================

# ── Core Flask ─────────────────────────────────────────────────────────
Flask==3.0.3
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
click==8.1.7
blinker==1.9.0

# ── Database ───────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36

# ── CORS + WebSocket ───────────────────────────────────────────────────
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
python-socketio==5.11.4
python-engineio==4.9.1
simple-websocket==1.1.0
eventlet==0.37.0
bidict==0.23.1

# ── LLM Providers (set whichever key you have in .env) ────────────────
# 1. ANTHROPIC_API_KEY  → Claude Haiku (console.anthropic.com — free credits)
# 2. GEMINI_API_KEY     → Gemini Flash (aistudio.google.com — 1M req/day free)
# 3. OPENAI_API_KEY     → GPT-4o-mini  (platform.openai.com — pay-per-use)
# 4. Rule-based fallback → zero cost, fully offline
anthropic==0.40.0
google-generativeai==0.8.3
openai==1.54.4
httpx==0.27.2
anyio==4.6.2
sniffio==1.3.1
h11==0.14.0

# ── ML / Data Science ─────────────────────────────────────────────────
scikit-learn==1.5.2
numpy==1.26.4
scipy==1.14.1
joblib==1.4.2
threadpoolctl==3.5.0

# ── Data Handling ──────────────────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0

# ── HTTP + Networking ──────────────────────────────────────────────────
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
tqdm==4.66.5

# ── Media & Files ──────────────────────────────────────────────────────
Pillow==10.4.0

# ── Config & Dev Tools ────────────────────────────────────────────────
python-dotenv==1.0.1
Faker==30.8.1

# ── Testing ────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
